getting there on day 10, part 1
Our initial approach was flawed, so we had to redesign our approach.
Now we make a copy of the grid, and iterate over it marking tiles as
visible/obscured from a given tile. In this way we work out what is
visible from a given tile. And then we simply make this calculation
once each for every tile in the input, and pick the biggest one as our
answer.

This works currently, but performs poorly, to the point where it is
impractical for the challenge input. So, from here, we will aim to
optimize this current design. Alternatively we could use a better
algorithm, that would avoid the duplication of work we have currently.
But if this current approach will perform good enough, then let's keep
it simple and go with that.