do day 7, part 2
After a painfully long amount of time debugging and scrutinizing the
code, we have eventually found the bug - we wrote

    sum (Set.map f set)

in order to compute the sum total of `f x` for all `x`s in set. But
this, of course, doesn't work - if `f x == f y` for any distinct `x`,
`y` in `set`, then only one of `f x` and `f y` will contribute to the
resulting sum. This is particularly insidious because, in small test
cases, it is easy to miss this edge case of `f` being non-injective.

At least, on the positive side, it is trivial to fix, as you can see in
the diff. It is also a forceful reminder about the behaviour of
`map :: (a -> b) -> Set a -> Set b`, and that you should only reach for
this function when you specifically want for duplicates to be merged.