do day 11, part 1
We repurpose our intcode machine for today, of course. But we also make
some small adjustments to it, such that it will be suitable as our
robots brain - we provide more control over waiting for inputs, and on
querying whether the robot is still running or not. We also do a
little house-keeping on the code in general.

Besides this, we build a `Robot` state machine that does what is
required - here the `zoom` function comes greatly in handy, to embed
the `Intcode` state machine inside this one. And we keep track of the
trail the robot has walked, so that we can answer part 1.