optimize day 11, part 2
Profiling revealed that we we're spending most of our time
in `inDirOf`. So rather than chaining together a bunch of smaller,
more-generic, functions, let's instead just write what we want
for this specific situation. That way we should avoid a bunch of
unnecessary calls we were making to `inDirOf`.

Indeed, our runtime has dropped from ~60s to ~4s.

(It should be possible to solve this in a way that keeps the
polymorphism of the functions, and the composability of having lots of
smaller functions. But this approach is simple and effective, and
that's all we need.)