do day 14, part 1
After some time spent playing around with different formulations, we
settled on a nice approach to part 1 as in this commit.

One thing to note, we went to some lengths to avoid interpreting any
integers as haskell `Int`s, because the challenge specifies that we are
working with 36-bit ints, and `Int` in haskell does not come with any
guarantee that it will be large enough for this. Indeed, we managed to
keep everything generic, and then at the end we specify that we will
use `Int64`s, more than large enough for 36 bits.