do day 22, part 2
What we missed in our initial pass is the rule that, in a recursive
game, it is normal behaviour for one player to lose by running out of
cards, like in part 1. This was not particularly clear to me from the
challenge description.

In any case, armed this fact, it is easy to enough to handle this rule
as we did in part 1. And then we're done.

This solution actually takes ~10s to run on my machine, much longer
than I expected. But I'm not sure how much recursion we are really
doing for the challenge input; perhaps it is a great deal. And this is
still quick-ish (< 15s) for advent of code. So we will take this
solution as good enough.