do day 17, part 2
This solution is slow (taking ~17s on my machine), but good enough for
AOC. All we do is modify the solution for part 1 to be parametric over
the dimension, and then parts 1 and 2 both fall out as special cases.

In terms of the actual details of the code, we have some type-level
machinery going on here - our main type is `V n Int` representing a
vector of `n` elements (represented at run-time as a (boxed) Vector
from `vector`). Then we need `Dim n` constraints in order to reify `n`
for functions like `withinOne`. And when we want to avoid writing a
proof that the dimensions match up, we can use the `V` constructor
explicitly.

Surprisingly though, most of the code needed almost no adjustment to be
parametric over the dimension. The trickiest part was adjusting the
extensions that are turned on so that the type-checker is happy.