do day 14 part 2
We move away from using `NESeq` and `Foldable1` so much, as for part
2 we would really want to have a `Seq` with at-least two elements and
a `Foldable2` class, which is getting a bit too cumbersome. Instead,
we sprinkle some `Maybe`s at the places where we really depend on the
length being a certain amount.

Besides this refactoring, the main idea behind our approach is not to
think of the problem as being an ever-growing list, but instead a map
from pairs to counts. Then, a rule such as `NC -> B` gets reinterpreted
as a function that maps `{NC: i}` to `{NB: i, BC: i}` where `i` here is
some `Int` (and we are using python-style notation for `Map`s here).

As long as we take some care to get the conversion functions right, we
can reframe our problem in terms of counters like this, and then it's
straightforward and quick to run the rules 40 times as needed.

Along the way we take advantage of `QuickCheck` to give us a lot of
confidence that our new approach produces the same results as our
naive version. (Indeed, it did not do so on our first attempt - which
QuickCheck helped us to catch - but it seems to do so now.)