do day 15 part 2
It is somewhat unsatisfactory, but our implementation from part 1 is
just about up to the task of solving part 2. Indeed, it is a bit slow
(~4s), but not so awful that it seems unacceptable.

So, there's not much to do for part 2, if we're just using the exact
same approach. However, as we elected to use `Matrix x y a` where
`x, y :: Nat` (and `a :: Type`), we are compelled to satisfy the type
checker when we quintuple the original input matrix from part 1. We
must prove that `n + n + n + n + n = 5 * n`, and that
`Dim n => Dim (5 * n)`. This is not trivial, but with a little
help from `unsafeCoerce` and some type-checker plugins we manage it.