do day 16 part 1
Pretty straightforward today, and something that Haskell is probably
well suited to. In any case, we employ the same technique we've used
before of doing multiple passes of parsing, so we start with

    parseHexChars :: Parser Text [Bit]

to go `Text -> [Bit]` and then we do (simplifying slightly)

    parsePacket :: Parser [Bit] (Packet Int)

to go `[Bit] -> Packet Int`. This works rather well, as `megaparsec`
is of course well suited to parsing the bits out from a bit-stream.

After doing all this parsing, summing up the version numbers is pretty
much trivial, as we made sure to keep all the metadata whilst parsing.