finish day 8
Mainly, we add alternate evaluators
which run the instructions and keep
track of what values have been
inserted into registers, to then be
able to compute the maximum.

For testing, we do some property
tests that our new optimized
tracer function `evalInstrsTraceMax`
produces the same results as the
original `evalInstrs`. Besides that,
we add the unit test from the
second part.