optimize day 05
We chose the data structure `Vector [Char]` as it should be quite
efficient for part 1 - we mainly want to push/pop from the from of the
stacks. But then for part 2, now we need to append a slice from one
stack onto the front of another. Now `[Char]` is looking more dubious,
so we switch it out for `Seq Char` which is O(1) for accessing the
ends and O(log n) for appending.

As it turns out, from some quick benchmarking, it doesn't seem to make
a difference after all here, as our example stacks are all very small.
Ah well, no harm in it.