finish day 12
For part 2 we do a similar process to Dijkstra, except that rather
than stopping once we find a particular node, we simply keep going
until we've explored the whole graph. Then we simply inspect our
recorded distances to pull out the shortest trail as needed.

At the same time we do some refactoring:

- we've already pulled in `extended-reals` for another day, which comes
  with convenient instances, so let's just use that instead.
- we provide a proper wrapper around `Matrix` which handles the
  `V2 a` to `(a, a)` conversion as needed for the unusual co-ordinate
  system that the `matrix` package uses.
- use better names for our `Algo` datatype from part 1.