optimize day 3, part 2
We skip the main loop from part 2 whenever we can. This is easy
enough - check the co-ordinates of the two positions being considered,
and if they are obviously far apart then we can skip it. A little
care is needed when subtracting `usize`s.

Having set this up, the runtime drops on my machine from ~13s
unoptimized (or ~1s release version) down to around .08s unoptimized.