do day 16, part 2
Sadly this part takes around 330ms (that is, .3s) to run on my
machine, which is around 30x slower than our other days have clocked
in at.

We could optimize this solution, as we clearly do a ton of
repeated work in calculating the same beams reflecting/splitting in the
same way. But it hardly seems worth the effort when the result still
runs in a fraction of a second. So we will accept this straightforward
solution.