update flake.nix
We have switched over to a different zig flake input that is supposed
to handle the wrapping of `/usr/bin/env` hardcoded into zig during the
build process.

Besides that, we have added in convenience outputs that can be used
like `nix run .#day-01` - this will compile day 1 in ReleaseSmall mode
(should be quicker than debug), point it at the input file via
environment variables, and then run it. It doesn't benefit from
caching so it's always slow, but it is nice to have the option.