refactor main/part1/part2 functions
Consider the following facts:

* All our main functions are (near) identical.
* We would like to be able to adjust how the output is printed
  in `main.zig`.
* With minor changes it would be possible to assign the type
  `fn (std.mem.Allocator, []const u8) !u64` to every `part1`
  and `part2` function so far.
* From a cursory inspection, it looks like every remaining part 1 and
  part 2 for days 10-25 could also be given the same type.

Given these factors, we've decided to do some light refactoring. The
result is that every part1 and part2 has the above type
`fn (std.mem.Allocator, []const u8) !u64`, every main function is a
lot simpler, and we can adjust the formatting in `main.zig` how we
please.

The downsides are that parts 1 and 2 are now forced to re-parse the
input doing duplicate work, and that there might be a possibility a
future day has a different output type. For now we'll accept these
tradeoffs.