finish 2025, day 2
To accommodate part 2, we rework `Doublet` into a more generalised
`Couplet`. Doublets can be used by replacing calls `Doublet.from(n)`
with `Couplet.from(2, n)`.

Armed with this change, solving part 2 is a simple matter of
checking Couplets of a few different sizes as needed and counting up
the hits as before.

We also use a nice trick with for-loops being expressions to tidy up
some code `var maybe: ?Doublet = null; ...` that is otherwise clunky.